Move CardTransformReward's RewardSynchronizer extensions into it's own file
