+++ Common/Rewards/CardTransformReward.cs
@@ -1,12 +1,19 @@
 using BaseLib.Abstracts;
 using BaseLib.Extensions;
 using BaseLib.Patches.Content;
+using MegaCrit.Sts2.Core.Commands;
 using MegaCrit.Sts2.Core.Entities.Players;
 using MegaCrit.Sts2.Core.Localization;
 using MegaCrit.Sts2.Core.Models.Cards;
 using MegaCrit.Sts2.Core.Rewards;
 using MegaCrit.Sts2.Core.Runs;
 using MegaCrit.Sts2.Core.Saves.Runs;
+using MegaCrit.Sts2.Core.CardSelection;
+using MegaCrit.Sts2.Core.Combat;
+using MegaCrit.Sts2.Core.Factories;
+using MegaCrit.Sts2.Core.Models;
+using MegaCrit.Sts2.Core.Multiplayer.Game;
+using MegaCrit.Sts2.Core.Nodes.CommonUi;
 
 namespace BaseLib.Common.Rewards;
 
@@ -102,5 +109,64 @@ public sealed class CardTransformReward(Player player) : CustomReward(player)
     {
         BaseLibMain.Logger.Info("Obtained card transformation from reward");
         return await RunManager.Instance.RewardSynchronizer.DoLocalCardTransform(Amount, true);
+    }
+}
+
+static class TransformRewardSynchronizerPatches
+{
+    /// <summary>
+    /// Method to handle transforming a card as a combat reward
+    /// </summary>
+    public static async Task<bool> DoLocalCardTransform(this RewardSynchronizer rewardSynchronizer, int amount = 1, bool upgrade = false)
+    {
+        CardTransformRewardMessage message = new CardTransformRewardMessage
+        {
+            Location = rewardSynchronizer._messageBuffer.CurrentLocation,
+            wasSkipped = false,
+            Upgrade = upgrade,
+            Amount = amount
+        };
+        BaseLibMain.Logger.Debug($"Transforming card for local player {rewardSynchronizer.LocalPlayer}");
+
+        CustomTargetedMessageWrapper.Send(message);
+        return await rewardSynchronizer.DoCardTransform(rewardSynchronizer.LocalPlayer, amount, upgrade);
+    }
+
+    /// <summary>
+    /// Transform a card for a specific player as a combat reward
+    /// </summary>
+    public static async Task<bool> DoCardTransform(this RewardSynchronizer rewardSynchronizer, Player player, int amount = 1, bool upgrade = false)
+    {
+        CardSelectorPrefs prefs = new CardSelectorPrefs(
+                upgrade
+                    ? CardSelectorPrefsExtensions.TransformAndUpgradeSelectionPrompt
+                    : CardSelectorPrefs.TransformSelectionPrompt,
+                1,
+                amount)
+        {
+            Cancelable = true,
+            RequireManualConfirmation = true
+        };
+
+        List<CardModel> cards = (await CardSelectCmd.FromDeckForTransformation(player, prefs)).ToList();
+
+        BaseLibMain.Logger.Debug($"Current combat state for transform rewards is: IsEnding={CombatManager.Instance.IsEnding}");
+        foreach (CardModel card in cards)
+        {
+            CardModel newCard = CardFactory.CreateRandomCardForTransform(
+                    card,
+                    isInCombat: false,
+                    player.RunState.Rng.Niche);
+
+            if (upgrade || card.IsUpgraded) // need a more robust handler for multi-upgrade at some point
+            {
+                CardCmd.Upgrade(newCard);
+            }
+
+            await CardCmd.Transform(card, newCard, CardPreviewStyle.GridLayout);
+            BaseLibMain.Logger.Debug($"Player {player.NetId} transformed {card.Id} in their deck into {newCard.Id}" + (upgrade ? " and upgraded it." : "."));
+        }
+
+        return cards.Count > 0;
     }
 }
